move util to super class
