+++ torch_frame/data/multi_embedding_tensor.py
@@ -206,32 +206,3 @@ class MultiEmbeddingTensor(_MultiTensor):
             # which is inconsistent with when dim=0
             offset = torch.LongTensor(offset_list)
             return MultiEmbeddingTensor(num_rows, num_cols, values, offset)
-
-    def _normalize_index(
-        self,
-        index: Union[int, Tensor],
-        dim: int,
-        is_slice_end: bool = False,
-    ) -> Union[int, Tensor]:
-        """Helper function to map negative indices to positive indices and
-        raise :obj:`IndexError` when necessary.
-
-        Args:
-            index: Union[int, Tensor]: Input :obj:`index` with potentially
-                negative elements.
-            is_slice_end (bool): Whether a given index (int) is slice or not.
-                If :obj:`True`, we have more lenient :obj:`IndexError`.
-                (default: :obj:`False`)
-        """
-        dim = self._normalize_dim(dim)
-        max_entries = self.num_rows if dim == 0 else self.num_cols
-        idx_name = "Row" if dim == 0 else "Col"
-        if isinstance(index, int):
-            if index < 0:
-                index = index + max_entries
-            if is_slice_end and index < 0 or index > max_entries:
-                raise IndexError(f"{idx_name} index out of bounds!")
-            elif (not is_slice_end) and (index < 0 or index >= max_entries):
-                raise IndexError(f"{idx_name} index out of bounds!")
-
-        return index
